requirements.txt
openai-whisper
setuptools-rust
fastapi
uvicorn
python-multipart
moviepy
youtube_dl
pytube
pydub
chardet